# [SPARK-16351][SQL] Avoid per-record type dispatch in JSON when writing