--- a/sql/core/src/main/scala/org/apache/spark/sql/Dataset.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/Dataset.scala
@@ -2478,6 +2478,7 @@ class Dataset[T] private[sql](
     val rowSchema = this.schema
     val rdd: RDD[String] = queryExecution.toRdd.mapPartitions { iter =>
       val writer = new CharArrayWriter()
+      // create the Generator without separator inserted between 2 records
       val gen = new JacksonGenerator(rowSchema, writer)
 
       new Iterator[String] {