@HyukjinKwon 2016-07-02
BTW, this approach can be applied to reading in JSON and CSV as well if this change is appropriate. I will work on them if this change is sensible and can be merged.