@HyukjinKwon 2016-07-04
@rxin and @yhuai Could you review this maybe?