@HyukjinKwon 2016-07-06
Actually, I resembled your codes @liancheng here. Would this change be sensible maybe?