@liancheng 2016-07-06
LGTM, but I'd like to let @yhuai to sign off. Thanks for working on this!