@HyukjinKwon 2016-07-08
Thanks @yhuai! I just addressed your comments.