@HyukjinKwon 2016-07-12
@yhuai Could you take another look?