@yhuai 2016-07-18
test this please