@yhuai 2016-07-18
LGTM pending jenkins.