@yhuai 2016-07-18
Merged to master.